advisor-gurus/backend/requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-JWT-Extended==4.4.4
Flask-CORS==3.0.10
Werkzeug==2.3.7
datetime==5.1.0
python-dotenv
gunicorn
